chore: extract state management into separate hook
